/cancel routing now takes a type (whitelisted) which is available to the cancellation flow PLUS refactored membership detail display into its own component
